src/requirements.txt
customtkinter
tkinter
pillow